Move the message dump commands to their own cog
