+++ cogs/role_commands.py
@@ -9,20 +9,6 @@ import voxelbotutils as utils
 class RoleCommands(utils.Cog):
 
     @utils.command()
-    @commands.has_permissions(manage_roles=True)
-    @commands.bot_has_permissions(send_messages=True)
-    async def copyrolepermissions(self, ctx:utils.Context, copy_role:discord.Role, paste_role:discord.Role):
-        """
-        Copies the permissions from one role onto another.
-        """
-
-        try:
-            await paste_role.edit(permissions=copy_role.permissions)
-        except discord.HTTPException:
-            return await ctx.send("Unable to edit role.")
-        return await ctx.send("Copied the permissions from {copy_role.mention} to {paste_role.mention}.", allowed_mentions=discord.AllowedMentions.none())
-
-    @utils.command()
     @commands.has_permissions(manage_messages=True)
     @commands.bot_has_permissions(send_messages=True)
     async def permissions(self, ctx:utils.Context, user:typing.Union[discord.Member, discord.Role], channel:typing.Optional[discord.TextChannel]):
@@ -58,36 +44,6 @@ class RoleCommands(utils.Cog):
             else:
                 output.append(f"Guild({green_circle if has_guild_permission else red_circle}) Channel({green_circle if has_channel_permission is True else red_circle if has_channel_permission is False else cross_emoji}) - **{permission_name}**")
         await ctx.send('\n'.join(output), allowed_mentions=discord.AllowedMentions(users=False, roles=False, everyone=False))
-
-    @utils.command()
-    @commands.has_permissions(manage_messages=True)
-    @commands.bot_has_permissions(send_messages=True)
-    async def dumpmessage(self, ctx:utils.Context, message:discord.Message):
-        """
-        Dumps a message out into chat.
-        """
-
-        await ctx.send(f'```\n{message.content}\n```')
-
-    @utils.command(aliases=['dumpmessageembed'])
-    @commands.has_permissions(manage_messages=True)
-    @commands.bot_has_permissions(send_messages=True)
-    async def dumpmessageembeds(self, ctx:utils.Context, message:discord.Message):
-        """
-        Dumps a message out into chat.
-        """
-
-        await ctx.send(f'```json\n{json.dumps([i.to_dict() for i in message.embeds], indent=2)}\n```')
-
-    @commands.command(cls=utils.Command)
-    @commands.has_permissions(manage_messages=True)
-    @commands.bot_has_permissions(send_messages=True)
-    async def dumpmessagelist(self, ctx:utils.Context, message:discord.Message):
-        """
-        Dumps a message out into chat.
-        """
-
-        await ctx.send(f'```py\n{[i.encode("unicode_escape").decode() for i in list(message.content)]}\n```')
 
 
 def setup(bot:utils.Bot):
